requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic[email]
passlib
pymysql
cryptography
pyjwt
alembic
python-multipart
slowapi